+++ NEWS.md
@@ -4,12 +4,16 @@
   - `extract_hive_partitions()` for extracting key value pairs from paths to hive-partitioned data files.
   - `is_hive_partitioned_path()` for checking if a path is hive-partitioned.
 * Added `read_target_file()` function for reading in individual target data files.
+* Added target data utilities:
+  - `get_target_task_id()` for extracting the name of the task ID(s) containing targets from the target metadata of a hub's config.
 * Added target data validation checks:
   - `check_target_file_name()`: that a hive-partitioned target data file name can be correctly parsed.
   - `check_target_dataset_unique()`: that a single unique target dataset exists for a given target type.
   - `check_target_dataset_file_ext_unique()`: that file(s) in a target dataset (e.g. `time-series` or `oracle-output`) share a single unique file extension.
   - `check_target_file_ext_valid()`: that the file extension of a single target data file is valid.
+  - `check_target_tbl_colnames()`: that column names of a target data file match expected column names for the target type.
 * Improved performance of `check_tbl_values_required()`.
+
 
 # hubValidations 0.11.0